# contracts/openapi.json
{
  "components": {
    "schemas": {
      "AccessLevel": {
        "enum": [
          "public",
          "course",
          "private"
        ],
        "title": "AccessLevel",
        "type": "string"
      },
...
      "Body_upload_and_parse_api_v1_resources_upload_post": {
        "properties": {
          "content_type": {
            "description": "文件类型：pdf/ppt/subtitle",
            "title": "Content Type",
            "type": "string"
          },
          "course_id": {
            "description": "课程 ID",
            "title": "Course Id",
            "type": "string"
          },
          "file": {
            "contentMediaType": "application/octet-stream",
            "description": "课程文件（PDF/PPT/SRT/VTT）",
            "title": "File",
            "type": "string"
          },
          "knowledge_point_ids": {
            "default": "",
            "description": "知识点 ID，逗号分隔",
            "title": "Knowledge Point Ids",
            "type": "string"
          },
          "language": {
            "default": "zh",
            "description": "语言：zh/en/bilingual",
            "title": "Language",
            "type": "string"
          },
          "title": {
            "description": "资料标题",
            "title": "Title",
            "type": "string"
          },
          "version": {
            "default": "v1",
            "description": "版本号",
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "file",
          "course_id",
          "title",
          "content_type"
        ],
        "title": "Body_upload_and_parse_api_v1_resources_upload_post",
        "type": "object"
      },
      "ChunkMetadata": {
        "additionalProperties": false,
        "properties": {
          "access_level": {
            "$ref": "#/components/schemas/AccessLevel"
          },
          "chapter": {
            "minLength": 1,
            "title": "Chapter",
            "type": "string"
          },
          "chunk_id": {
            "minLength": 1,
            "title": "Chunk Id",
            "type": "string"
          },
          "content": {
            "minLength": 1,
            "title": "Content",
            "type": "string"
          },
          "content_type": {
            "$ref": "#/components/schemas/ContentType"
          },
          "extracted_content_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ContentType"
              },
              {
                "type": "null"
              }
            ]
          },
          "knowledge_point_ids": {
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Knowledge Point Ids",
            "type": "array"
          },
          "language": {
            "$ref": "#/components/schemas/Language"
          },
          "page_end": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page End"
          },
          "page_start": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Start"
          },
          "parse_status": {
            "$ref": "#/components/schemas/ParseStatus"
          },
          "resource_id": {
            "minLength": 1,
            "title": "Resource Id",
            "type": "string"
          },
          "source_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Path"
          },
          "source_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Url"
          },
          "title": {
            "minLength": 1,
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "version": {
            "minLength": 1,
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "chunk_id",
          "resource_id",
          "knowledge_point_ids",
          "title",
          "content",
          "chapter",
          "language",
          "content_type",
          "version",
          "access_level",
          "parse_status",
          "updated_at"
        ],
        "title": "ChunkMetadata",
        "type": "object"
      },
...
      "ContentType": {
        "enum": [
          "pdf",
          "ppt",
          "subtitle",
          "text",
          "rfc",
          "table",
          "formula",
          "image",
          "diagram",
          "other"
        ],
        "title": "ContentType",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "KnowledgeBaseChangeStatus": {
        "enum": [
          "pending",
          "in_progress",
          "done",
          "wont_fix"
        ],
        "title": "KnowledgeBaseChangeStatus",
        "type": "string"
      },
      "KnowledgeBaseChangeTask": {
        "additionalProperties": false,
        "properties": {
          "change_id": {
            "minLength": 1,
            "title": "Change Id",
            "type": "string"
          },
          "course_id": {
            "minLength": 1,
            "title": "Course Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "feedback_id": {
            "minLength": 1,
            "title": "Feedback Id",
            "type": "string"
          },
          "handler": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Handler"
          },
          "question": {
            "minLength": 1,
            "title": "Question",
            "type": "string"
          },
          "reason": {
            "minLength": 1,
            "title": "Reason",
            "type": "string"
          },
          "request_id": {
            "minLength": 1,
            "title": "Request Id",
            "type": "string"
          },
          "resolution_notes": {
            "default": "",
            "title": "Resolution Notes",
            "type": "string"
          },
          "resource_ids": {
            "description": "本次优化实际新增或修订的课程资料标识。",
            "items": {
              "type": "string"
            },
            "title": "Resource Ids",
            "type": "array"
          },
          "status": {
            "$ref": "#/components/schemas/KnowledgeBaseChangeStatus"
          },
          "suggested_action": {
            "minLength": 1,
            "title": "Suggested Action",
            "type": "string"
          },
          "updated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "required": [
          "change_id",
          "feedback_id",
          "course_id",
          "request_id",
          "question",
          "reason",
          "suggested_action",
          "status",
          "created_at"
        ],
        "title": "KnowledgeBaseChangeTask",
        "type": "object"
      },
      "KnowledgeBaseChangeUpdateRequest": {
        "additionalProperties": false,
        "description": "推进或关闭知识库变更任务，构成“审核无效回答 → 优化知识库”的收尾环节。",
        "properties": {
          "handler": {
            "minLength": 1,
            "title": "Handler",
            "type": "string"
          },
          "resolution_notes": {
            "default": "",
            "maxLength": 2000,
            "title": "Resolution Notes",
            "type": "string"
          },
          "resource_ids": {
            "items": {
              "type": "string"
            },
            "title": "Resource Ids",
            "type": "array"
          },
          "status": {
            "$ref": "#/components/schemas/KnowledgeBaseChangeStatus"
          }
        },
        "required": [
          "status",
          "handler"
        ],
        "title": "KnowledgeBaseChangeUpdateRequest",
        "type": "object"
      },
...
      "Language": {
        "enum": [
          "zh",
          "en",
          "bilingual"
        ],
        "title": "Language",
        "type": "string"
      },
...
      "ParseStatus": {
        "enum": [
          "pending",
          "parsed",
          "failed"
        ],
        "title": "ParseStatus",
        "type": "string"
      },
...
      "ResourceImportRequest": {
        "additionalProperties": false,
        "properties": {
          "content_type": {
            "$ref": "#/components/schemas/ContentType"
          },
          "course_id": {
            "minLength": 1,
            "title": "Course Id",
            "type": "string"
          },
          "language": {
            "$ref": "#/components/schemas/Language"
          },
          "resource_type": {
            "$ref": "#/components/schemas/ResourceType"
          },
          "source_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Url"
          },
          "title": {
            "minLength": 1,
            "title": "Title",
            "type": "string"
          },
          "version": {
            "minLength": 1,
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "course_id",
          "resource_type",
          "title",
          "version",
          "language",
          "content_type"
        ],
        "title": "ResourceImportRequest",
        "type": "object"
      },
      "ResourceImportResponse": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "metadata": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ChunkMetadata"
              },
              {
                "type": "null"
              }
            ]
          },
          "resource_id": {
            "minLength": 1,
            "title": "Resource Id",
            "type": "string"
          },
          "sync_status": {
            "$ref": "#/components/schemas/SyncStatus"
          }
        },
        "required": [
          "resource_id",
          "sync_status"
        ],
        "title": "ResourceImportResponse",
        "type": "object"
      },
      "ResourceSummary": {
        "additionalProperties": false,
        "description": "已导入资料的概览，用于确认解析与入库结果。",
        "properties": {
          "chunk_count": {
            "minimum": 1.0,
            "title": "Chunk Count",
            "type": "integer"
          },
          "content_type": {
            "$ref": "#/components/schemas/ContentType"
          },
          "course_id": {
            "minLength": 1,
            "title": "Course Id",
            "type": "string"
          },
          "language": {
            "$ref": "#/components/schemas/Language"
          },
          "resource_id": {
            "minLength": 1,
            "title": "Resource Id",
            "type": "string"
          },
          "title": {
            "minLength": 1,
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "version": {
            "minLength": 1,
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "resource_id",
          "course_id",
          "title",
          "chunk_count",
          "language",
          "content_type",
          "version",
          "updated_at"
        ],
        "title": "ResourceSummary",
        "type": "object"
      },
      "ResourceType": {
        "enum": [
          "textbook",
          "slide",
          "lab_guide",
          "video_subtitle",
          "rfc",
          "other"
        ],
        "title": "ResourceType",
        "type": "string"
      },
      "ResourceUploadResponse": {
        "additionalProperties": false,
        "properties": {
          "chunk_count": {
            "minimum": 0.0,
            "title": "Chunk Count",
            "type": "integer"
          },
          "chunks": {
            "items": {
              "$ref": "#/components/schemas/ChunkMetadata"
            },
            "title": "Chunks",
            "type": "array"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "filename": {
            "minLength": 1,
            "title": "Filename",
            "type": "string"
          },
          "indexed_chunks": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "向量库在本次导入后的 Chunk 总数；为空表示未执行向量化。",
            "title": "Indexed Chunks"
          },
          "parse_status": {
            "$ref": "#/components/schemas/ParseStatus"
          },
          "resource_id": {
            "minLength": 1,
            "title": "Resource Id",
            "type": "string"
          }
        },
        "required": [
          "resource_id",
          "filename",
          "chunk_count",
          "parse_status"
        ],
        "title": "ResourceUploadResponse",
        "type": "object"
      },
...
      "SyncStatus": {
        "enum": [
          "accepted",
          "completed",
          "failed"
        ],
        "title": "SyncStatus",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "完全独立于 Canvas 的离线优先项目骨架。",
    "title": "计算机网络 AI 助教 API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/knowledge-base/changes": {
      "get": {
        "operationId": "list_changes_api_v1_knowledge_base_changes_get",
        "parameters": [
          {
            "in": "query",
            "name": "course_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Course Id"
            }
          },
          {
            "in": "query",
            "name": "feedback_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Feedback Id"
            }
          },
          {
            "in": "query",
            "name": "change_status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/KnowledgeBaseChangeStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Change Status"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/KnowledgeBaseChangeTask"
                  },
                  "title": "Response List Changes Api V1 Knowledge Base Changes Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Changes",
        "tags": [
          "knowledge-base"
        ]
      }
    },
    "/api/v1/knowledge-base/changes/{change_id}": {
      "get": {
        "operationId": "get_change_api_v1_knowledge_base_changes__change_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "change_id",
            "required": true,
            "schema": {
              "title": "Change Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeBaseChangeTask"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Change",
        "tags": [
          "knowledge-base"
        ]
      },
      "patch": {
        "description": "推进或关闭知识库变更任务，完成知识库持续优化的闭环。",
        "operationId": "update_change_api_v1_knowledge_base_changes__change_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "change_id",
            "required": true,
            "schema": {
              "title": "Change Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KnowledgeBaseChangeUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeBaseChangeTask"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Change",
        "tags": [
          "knowledge-base"
        ]
      }
    },
...
    "/api/v1/resources": {
      "get": {
        "description": "列出已导入的课程资料，便于确认解析结果是否入库。",
        "operationId": "list_resources_api_v1_resources_get",
        "parameters": [
          {
            "in": "query",
            "name": "course_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Course Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ResourceSummary"
                  },
                  "title": "Response List Resources Api V1 Resources Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Resources",
        "tags": [
          "resources"
        ]
      }
    },
    "/api/v1/resources/import": {
      "post": {
        "operationId": "import_resource_api_v1_resources_import_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResourceImportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResourceImportResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Import Resource",
        "tags": [
          "resources"
        ]
      }
    },
    "/api/v1/resources/upload": {
      "post": {
        "description": "上传课程文件，自动解析、切块、向量化入库。\n\n支持 PDF、PPT（.pptx）、SRT 和 VTT 字幕文件。\n解析后的 Chunk 会存入 ChunkStore，并增量写入向量库，\n使新导入的资料可以立即被问答检索和引用。",
        "operationId": "upload_and_parse_api_v1_resources_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_and_parse_api_v1_resources_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResourceUploadResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload And Parse",
        "tags": [
          "resources"
        ]
      }
    },
    "/api/v1/resources/{resource_id}/chunks": {
      "get": {
        "description": "读取某份资料解析出的全部 Chunk，用于人工核对切块与引用定位。",
        "operationId": "list_resource_chunks_api_v1_resources__resource_id__chunks_get",
        "parameters": [
          {
            "in": "path",
            "name": "resource_id",
            "required": true,
            "schema": {
              "title": "Resource Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ChunkMetadata"
                  },
                  "title": "Response List Resource Chunks Api V1 Resources  Resource Id  Chunks Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Resource Chunks",
        "tags": [
          "resources"
        ]
      }
    },
...
  }
}